requirements.txt
beautifulsoup4
requests
google-generativeai
dotenv
selenium
pillow
jinja2
pdfkit